requirements.txt
torch
torchvision
tensorflow
numpy
Pillow
matplotlib
opencv-python
scikit-image
discord.py
